+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support to print data types Quaternion, EulerAngles, WorldPosition and 
   WorldTransform.
+- scroll to property's offset in MEMORY view when hovering it in PROPERTIES 
+  view.
 
 ------------------------